#91 [feat] 추천 알고리즘 로직 제작

+++ src/main/java/org/sopt/certi_server/domain/admin/dto/request/CertificationJobCreateRequest.java
@@ -1,4 +1,4 @@
 package org.sopt.certi_server.domain.admin.dto.request;
 
-public record CertificationJobCreateRequest(String certificationName, String jobName, float weight) {
+public record CertificationJobCreateRequest(String certificationName, String jobName, double weight) {
 }
