+++ src/main/java/org/sopt/certi_server/domain/admin/dto/request/CertificationMajorCreateRequest.java
@@ -1,4 +1,4 @@
 package org.sopt.certi_server.domain.admin.dto.request;
 
-public record CertificationMajorCreateRequest(String certificationName, String majorName, float weight) {
+public record CertificationMajorCreateRequest(String certificationName, String majorName, double weight) {
 }
